refactor(frontend): improve code ml-339
