+++ apps/frontend/src/libs/components/meeting-pdf/libs/components/render-markdown.tsx
@@ -5,8 +5,8 @@ import {
 	type PhrasingContent,
 } from "~/libs/types/types.js";
 
-import { styles } from "../helpers/meeting-pdf.styles.js";
 import { parseMarkdown } from "../helpers/parse-markdown.js";
+import { styles } from "../styles/meeting-pdf.styles.js";
 import { renderNode } from "./render-node.js";
 
 type MdNode = BlockContent | ListItem | PhrasingContent;
